deployments/docker/academy/requirements.txt
social-auth-app-django==3.1.0
Django==3.0.7
ipython==8.4.0
# mysqlclient==1.4.6
PyVimeo==1.1.0
requests==2.25.1
kubernetes==17.17.0
django-filter==2.4.0
django-paypal==1.1.1
django-recaptcha==2.0.6
django-markdownify==0.9.1
paypal==1.2.5
django-mysql==4.5.0
django-extensions==3.1.5
jsonfield==3.1.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
openai
celery==5.3.5
redis==5.0.1
